+++ backend/src/common/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './slug.util';
+export * from './test-helpers.util';